requirements.txt
pynput==1.7.6
pyserial==3.5
six==1.16.0
